+++ test/components/slope/test_slope_analysis.py
@@ -8,7 +8,7 @@ from shapely.geometry.multilinestring import MultiLineString
 from walkability.components.slope.slope_analysis import get_slope
 
 
-def test_get_slope(global_aoi, responses_mock, operator):
+def test_get_slope(global_aoi, responses_mock, default_ors_settings):
     responses_mock.post(
         'https://api.openrouteservice.org/elevation/line',
         json={
@@ -41,7 +41,7 @@ def test_get_slope(global_aoi, responses_mock, operator):
         ],
         crs='EPSG:4326',
     )
-    computed_slope = get_slope(aoi=global_aoi, paths=paths, ors_client=operator.ors_settings.client)
+    computed_slope = get_slope(aoi=global_aoi, paths=paths, ors_settings=default_ors_settings)
 
     expected_naturalness = gpd.GeoDataFrame(
         index=[1],
@@ -55,7 +55,7 @@ def test_get_slope(global_aoi, responses_mock, operator):
     assert_geodataframe_equal(computed_slope, expected_naturalness, check_like=True)
 
 
-def test_get_negative_slope(global_aoi, responses_mock, operator):
+def test_get_negative_slope(global_aoi, responses_mock, default_ors_settings):
     responses_mock.post(
         'https://api.openrouteservice.org/elevation/line',
         json={
@@ -91,7 +91,7 @@ def test_get_negative_slope(global_aoi, responses_mock, operator):
         data={'@osmId': ['way/a', 'way/b']},
         crs='EPSG:4326',
     )
-    computed_slope = get_slope(aoi=global_aoi, paths=paths, ors_client=operator.ors_settings.client)
+    computed_slope = get_slope(aoi=global_aoi, paths=paths, ors_settings=default_ors_settings)
 
     expected_naturalness = gpd.GeoDataFrame(
         index=[1, 2],
@@ -106,7 +106,7 @@ def test_get_negative_slope(global_aoi, responses_mock, operator):
     assert_geodataframe_equal(computed_slope, expected_naturalness, check_like=True)
 
 
-def test_get_duplicate_slope(global_aoi, responses_mock, operator):
+def test_get_duplicate_slope(global_aoi, responses_mock, default_ors_settings):
     responses_mock.post(
         'https://api.openrouteservice.org/elevation/line',
         json={
@@ -141,7 +141,7 @@ def test_get_duplicate_slope(global_aoi, responses_mock, operator):
         data={'@osmId': ['way/a', 'way/b']},
         crs='EPSG:4326',
     )
-    computed_slope = get_slope(aoi=global_aoi, paths=paths, ors_client=operator.ors_settings.client)
+    computed_slope = get_slope(aoi=global_aoi, paths=paths, ors_settings=default_ors_settings)
 
     expected_slope = gpd.GeoDataFrame(
         index=[1, 2],
@@ -156,7 +156,7 @@ def test_get_duplicate_slope(global_aoi, responses_mock, operator):
     assert_geodataframe_equal(computed_slope, expected_slope, check_like=True)
 
 
-def test_slope_matching_according_to_ors_precision(responses_mock, global_aoi, operator):
+def test_slope_matching_according_to_ors_precision(responses_mock, global_aoi, default_ors_settings):
     responses_mock.post(
         'https://api.openrouteservice.org/elevation/line',
         json={
@@ -189,7 +189,7 @@ def test_slope_matching_according_to_ors_precision(responses_mock, global_aoi, o
         data={'@osmId': ['way/a']},
         crs='EPSG:4326',
     )
-    computed_slope = get_slope(aoi=global_aoi, paths=paths, ors_client=operator.ors_settings.client)
+    computed_slope = get_slope(aoi=global_aoi, paths=paths, ors_settings=default_ors_settings)
 
     expected_naturalness = gpd.GeoDataFrame(
         index=[1],
@@ -203,7 +203,7 @@ def test_slope_matching_according_to_ors_precision(responses_mock, global_aoi, o
     assert_geodataframe_equal(computed_slope, expected_naturalness, check_like=True)
 
 
-def test_get_large_amount_of_slopes(global_aoi, responses_mock, operator):
+def test_get_large_amount_of_slopes(global_aoi, responses_mock, default_ors_settings):
     responses_mock.post(
         'https://api.openrouteservice.org/elevation/line',
         json={
@@ -247,9 +247,7 @@ def test_get_large_amount_of_slopes(global_aoi, responses_mock, operator):
 
     geoms = [LineString([[0.0, 0.0], [1.0, 1.0]]), LineString([[1.0, 1.0], [2.0, 2.0]])]
     paths = gpd.GeoDataFrame(geometry=geoms, crs='EPSG:4326', data={'@osmId': ['way/a', 'way/b']})
-    computed_slope = get_slope(
-        aoi=global_aoi, paths=paths, ors_client=operator.ors_settings.client, request_chunk_size=2
-    )
+    computed_slope = get_slope(aoi=global_aoi, paths=paths, ors_settings=default_ors_settings, request_chunk_size=2)
 
     expected_slope = gpd.GeoDataFrame(
         geometry=geoms,
@@ -260,7 +258,7 @@ def test_get_large_amount_of_slopes(global_aoi, responses_mock, operator):
     assert_geodataframe_equal(computed_slope, expected_slope, check_like=True)
 
 
-def test_slope_for_multipart_geom(responses_mock, global_aoi, operator):
+def test_slope_for_multipart_geom(responses_mock, global_aoi, default_ors_settings):
     responses_mock.post(
         'https://api.openrouteservice.org/elevation/line',
         json={
@@ -293,7 +291,7 @@ def test_slope_for_multipart_geom(responses_mock, global_aoi, operator):
         data={'@osmId': ['way/a']},
         crs='EPSG:4326',
     )
-    computed_slope = get_slope(aoi=global_aoi, paths=paths, ors_client=operator.ors_settings.client)
+    computed_slope = get_slope(aoi=global_aoi, paths=paths, ors_settings=default_ors_settings)
 
     expected_naturalness = gpd.GeoDataFrame(
         index=[1],