+++ test/core/test_operator_worker.py
@@ -10,41 +10,7 @@ from walkability.components.utils.misc import (
 )
 
 
-def test_get_buffered_line_paths(operator, expected_compute_input, default_aoi, ohsome_api):
-    line_geom = shapely.LineString([(12.3, 48.22), (12.3, 48.2205), (12.3005, 48.22)])
-
-    expected_lines = gpd.GeoDataFrame(
-        data={
-            '@osmId': ['way/171574582'],
-            'category': [PathCategory.DESIGNATED],
-            'rating': [1.0],
-            'quality': [PavementQuality.UNKNOWN],
-            'quality_rating': [None],
-            'smoothness': [SmoothnessCategory.UNKNOWN],
-            'smoothness_rating': [None],
-            'surface': [SurfaceType.UNKNOWN],
-            'surface_rating': [None],
-            'geometry': [line_geom],
-            '@other_tags': [{'highway': 'pedestrian'}],
-        },
-        crs='EPSG:4326',
-    )
-
-    _, computed_lines_buffered, _ = operator._get_paths(
-        aoi=default_aoi,
-        max_walking_distance=expected_compute_input.max_walking_distance,
-    )
-
-    testing.assert_geodataframe_equal(
-        computed_lines_buffered,
-        expected_lines,
-        check_like=True,
-        check_geom_type=True,
-        check_less_precise=True,
-    )
-
-
-def test_get_line_paths(operator, expected_compute_input, ohsome_api):
+def test_get_line_paths(operator, ohsome_api):
     line_geom = [
         shapely.LineString([(12.3, 48.22), (12.3, 48.2205)]),
         shapely.LineString([(12.3, 48.2205), (12.3005, 48.22)]),
@@ -67,7 +33,7 @@ def test_get_line_paths(operator, expected_compute_input, ohsome_api):
         crs='EPSG:4326',
     )
 
-    computed_lines, _, _ = operator._get_paths(
+    computed_lines, _ = operator._get_paths(
         aoi=shapely.MultiPolygon(
             polygons=[
                 [
@@ -80,8 +46,7 @@ def test_get_line_paths(operator, expected_compute_input, ohsome_api):
                     ]
                 ]
             ]
-        ),
-        max_walking_distance=expected_compute_input.max_walking_distance,
+        )
     )
 
     testing.assert_geodataframe_equal(
@@ -93,7 +58,7 @@ def test_get_line_paths(operator, expected_compute_input, ohsome_api):
     )
 
 
-def test_get_polygon_paths(operator, expected_compute_input, default_aoi, ohsome_api):
+def test_get_polygon_paths(operator, default_aoi, ohsome_api):
     polygon_geom = shapely.Polygon(((12.3, 48.22), (12.3, 48.2205), (12.3005, 48.22), (12.3, 48.22)))
 
     expected_polygons = gpd.GeoDataFrame(
@@ -113,10 +78,7 @@ def test_get_polygon_paths(operator, expected_compute_input, default_aoi, ohsome
         crs='EPSG:4326',
     )
 
-    _, _, computed_polygons = operator._get_paths(
-        aoi=default_aoi,
-        max_walking_distance=expected_compute_input.max_walking_distance,
-    )
+    _, computed_polygons = operator._get_paths(aoi=default_aoi)
 
     testing.assert_geodataframe_equal(
         computed_polygons,