+++ test/test_plugin.py
@@ -21,22 +21,46 @@ def mock_detour_factor_calculation(expected_detour_factors):
         yield detour_factors
 
 
-def test_plugin_compute_request(
+def test_plugin_compute_request_minimal(operator, default_aoi, default_aoi_properties, compute_resources, ohsome_api):
+    computed_artifacts = operator.compute(
+        resources=compute_resources,
+        aoi=default_aoi,
+        aoi_properties=default_aoi_properties,
+        params=ComputeInputWalkability(),
+    )
+
+    assert len(computed_artifacts) == 6
+    for artifact in computed_artifacts:
+        assert isinstance(artifact, _Artifact)
+
+
+def test_plugin_compute_request_all_optionals(
     operator,
     expected_compute_input,
     default_aoi,
     default_aoi_properties,
     compute_resources,
     ohsome_api,
+    ors_isochrone_api,
 ):
-    with open('test/resources/ohsome_admin_response.geojson', 'r') as admin_file:
+    with (
+        open('test/resources/ohsome_admin_response.geojson', 'r') as admin_file,
+        open('test/resources/ohsome_drinking_water.geojson', 'r') as drinking_water,
+    ):
         admin_body = admin_file.read()
+        drinking_water_body = drinking_water.read()
     ohsome_api.post(
         'https://api.ohsome.org/v1/elements/geometry',
         body=admin_body,
         match=[filter_start_matcher('geometry:polygon and boundary')],
     )
+    ohsome_api.post(
+        'https://api.ohsome.org/v1/elements/centroid',
+        body=drinking_water_body,
+    )
 
+    expected_compute_input = expected_compute_input.model_copy(deep=True)
+    expected_compute_input.optional_indicators = {e for e in WalkabilityIndicators}
     computed_artifacts = operator.compute(
         resources=compute_resources,
         aoi=default_aoi,
@@ -44,42 +68,6 @@ def test_plugin_compute_request(
         params=expected_compute_input,
     )
 
-    assert len(computed_artifacts) == 8
+    assert len(computed_artifacts) == 12
     for artifact in computed_artifacts:
         assert isinstance(artifact, _Artifact)
-
-
-def test_plugin_compute_request_empty(operator, default_aoi, default_aoi_properties, compute_resources, ohsome_api):
-    computed_artifacts = operator.compute(
-        resources=compute_resources,
-        aoi=default_aoi,
-        aoi_properties=default_aoi_properties,
-        params=ComputeInputWalkability(walkable_categories_selection=set()),
-    )
-
-    assert len(computed_artifacts) == 6
-    for artifact in computed_artifacts:
-        assert isinstance(artifact, _Artifact) or artifact is None
-
-
-def test_plugin_compute_request_with_only_one_optional_indicator(
-    operator,
-    default_aoi,
-    default_aoi_properties,
-    compute_resources,
-    ohsome_api,
-    ors_elevation_api,
-    mock_detour_factor_calculation,
-):
-    computed_artifacts = operator.compute(
-        resources=compute_resources,
-        aoi=default_aoi,
-        aoi_properties=default_aoi_properties,
-        params=ComputeInputWalkability(
-            optional_indicators={WalkabilityIndicators.SLOPE, WalkabilityIndicators.DETOURS}
-        ),
-    )
-
-    assert len(computed_artifacts) == 10
-    for artifact in computed_artifacts:
-        assert isinstance(artifact, _Artifact) or artifact is None