+++ walkability/components/slope/slope_analysis.py
@@ -4,14 +4,14 @@ from typing import Tuple
 
 import geopandas as gpd
 import numpy as np
-import openrouteservice
 import pandas as pd
 import shapely
 from climatoology.base.baseoperator import _Artifact
 from climatoology.base.computation import ComputationResources
 
 from walkability.components.slope.slope_artifacts import build_slope_artifact
-from walkability.components.utils.geometry import ORS_COORDINATE_PRECISION, get_utm_zone
+from walkability.components.utils.geometry import get_utm_zone
+from walkability.components.utils.ors_settings import ORSSettings
 
 log = logging.getLogger(__name__)
 
@@ -19,11 +19,11 @@ log = logging.getLogger(__name__)
 def slope_analysis(
     line_paths: gpd.GeoDataFrame,
     aoi: shapely.MultiPolygon,
-    ors_client: openrouteservice.Client,
+    ors_settings: ORSSettings,
     resources: ComputationResources,
 ) -> Tuple[_Artifact, gpd.GeoDataFrame]:
     log.info('Computing slope')
-    slope = get_slope(paths=line_paths, aoi=aoi, ors_client=ors_client)
+    slope = get_slope(paths=line_paths, aoi=aoi, ors_settings=ors_settings)
     slope_artifact = build_slope_artifact(slope=slope, resources=resources)
     log.info('Finished computing slope')
 
@@ -33,12 +33,12 @@ def slope_analysis(
 def get_slope(
     paths: gpd.GeoDataFrame,
     aoi: shapely.MultiPolygon,
-    ors_client: openrouteservice.Client,
+    ors_settings: ORSSettings,
     request_chunk_size: int = 2000,
 ) -> gpd.GeoDataFrame:
     """Retrieve the slope of paths.
 
-    :param ors_client:
+    :param ors_settings:
     :param paths:
     :param aoi:
     :param request_chunk_size: Maximum number of elevation to be requested from the server. The server has a limit set that must be respected.
@@ -58,7 +58,7 @@ def get_slope(
     # PERF: do parallel calls (using async?)
     for chunk in np.array_split(points, num_chunks):
         polyline = list(zip(chunk.x, chunk.y))
-        response = ors_client.elevation_line(
+        response = ors_settings.client.elevation_line(
             format_in='polyline', format_out='polyline', dataset='srtm', geometry=polyline
         )
 
@@ -67,8 +67,10 @@ def get_slope(
             point = shapely.Point(coord[0:2])
             ele = coord[2]
 
-            start_point_match = paths['start_point'].geom_equals_exact(point, tolerance=ORS_COORDINATE_PRECISION)
-            end_point_match = paths['end_point'].geom_equals_exact(point, tolerance=ORS_COORDINATE_PRECISION)
+            start_point_match = paths['start_point'].geom_equals_exact(
+                point, tolerance=ors_settings.coordinate_precision
+            )
+            end_point_match = paths['end_point'].geom_equals_exact(point, tolerance=ors_settings.coordinate_precision)
 
             paths.loc[start_point_match, 'start_ele'] = ele
             paths.loc[end_point_match, 'end_ele'] = ele